requirements.txt
gradio==5.35.0
strands-agents==0.2.0
strands-agents-tools==0.1.8
asyncio